HACK: extcon: max8997: identify 619 kOhm jig as uart
max8997 identifies a 619 kOhm debug cable as an audio dock, while most
other samsung devices identifies it as a uart jig. This causes uart
output to stop as muic is probed. For max8997 it seems to be required
to use a 523 kOhm jig instead.

I do not have any dock, and do not have the parts to make a new cable
with 523 kOhm, so hack driver to identify a 619 kOhm device as an uart
jig as well.